textailes-api/requirements.txt
aniso8601==10.0.1
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
Authlib==1.6.5
avro==1.12.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
confluent-kafka==2.12.0
cryptography==46.0.2
exceptiongroup==1.3.0
fastavro==1.12.1
flasgger==0.9.7.1
Flask==3.1.2
Flask-RESTful==0.3.10
flask-swagger-ui==5.21.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
minio==7.2.18
mistune==3.1.4
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
